06-12-13 [UPD] delete adjustments

+++ dirIterator/Filters/Filter.php
@@ -5,11 +5,14 @@ namespace dirIterator\Filters;
 abstract class Filter extends \FilterIterator {
 
     protected $_it;
+    protected static $increment = 0;
 
     abstract public function count();
 
     public function delete() {
-        return unlink($this->_it->getPathName());
+        if (!$this->_it->isDir()) {
+            return unlink($this->_it->getPathName());
+        }
     }
 
 }
